feat: shift + arrow_up/down for bulk selection
